+++ src/find-options/FindOneOptions.ts
@@ -32,9 +32,8 @@ export interface FindOneOptions<Entity> {
     order?: { [P in keyof Entity]?: "ASC"|"DESC"|1|-1 };
 
     /**
-     * If sets to true then loads all relation ids of the entity and maps them into relation values (not relation objects).
-     * If array of strings is given then loads only relation ids of the given properties.
+     * Enables or disables query result caching.
      */
-    loadRelationIds?: boolean|{ relations?: string[], disableMixedMap?: boolean }; // todo: extract options into separate interface, reuse
+    cache?: boolean | number | { id: any, milliseconds: number };
 
 }