+++ src/repository/BaseEntity.ts
@@ -22,6 +22,7 @@ export class BaseEntity {
     /**
      * Connection used in all static methods of the BaseEntity.
      */
+    // @ts-ignore: Unused variable which is actually used
     private static usedConnection?: Connection;
 
     // -------------------------------------------------------------------------
@@ -104,8 +105,20 @@ export class BaseEntity {
     /**
      * Creates a new entity instance.
      */
-    static create<T extends BaseEntity>(this: ObjectType<T>): T {
-        return (this as any).getRepository().create();
+    static create<T extends BaseEntity>(this: ObjectType<T>): T;
+    /**
+     * Creates a new entities and copies all entity properties from given objects into their new entities.
+     * Note that it copies only properties that present in entity schema.
+     */
+    static create<T extends BaseEntity>(this: ObjectType<T>, entityLikeArray: DeepPartial<T>[]): T;
+    /**
+     * Creates a new entity instance and copies all entity properties from this object into a new entity.
+     * Note that it copies only properties that present in entity schema.
+     */
+    static create<T extends BaseEntity>(this: ObjectType<T>, entityLike: DeepPartial<T>): T;
+   
+    static create<T extends BaseEntity>(this: ObjectType<T>, entityOrEntities?: any): T {
+        return (this as any).getRepository().create(entityOrEntities);
     }
 
     /**