+++ test/functional/database-schema/column-types/postgres/column-types-postgres.ts
@@ -48,6 +48,7 @@ describe("database schema > column types > postgres", () => {
         post.varchar = "This is varchar";
         post.characterVarying = "This is character varying";
         post.text = "This is text";
+        post.citext = "This is text";
         post.bytea = new Buffer("This is bytea");
         post.date = "2017-06-21";
         post.interval = "1 year 2 months 3 days 4 hours 5 minutes 6 seconds";
@@ -104,6 +105,7 @@ describe("database schema > column types > postgres", () => {
         loadedPost.varchar.should.be.equal(post.varchar);
         loadedPost.characterVarying.should.be.equal(post.characterVarying);
         loadedPost.text.should.be.equal(post.text);
+        loadedPost.citext.should.be.equal("this is text");
         loadedPost.bytea.toString().should.be.equal(post.bytea.toString());
         loadedPost.date.should.be.equal(post.date);
         loadedPost.interval.years.should.be.equal(1);
@@ -165,6 +167,7 @@ describe("database schema > column types > postgres", () => {
         table!.findColumnByName("varchar")!.type.should.be.equal("character varying");
         table!.findColumnByName("characterVarying")!.type.should.be.equal("character varying");
         table!.findColumnByName("text")!.type.should.be.equal("text");
+        table!.findColumnByName("citext")!.type.should.be.equal("citext");
         table!.findColumnByName("bytea")!.type.should.be.equal("bytea");
         table!.findColumnByName("date")!.type.should.be.equal("date");
         table!.findColumnByName("interval")!.type.should.be.equal("interval");