+++ test/integration/sample4-many-to-many.ts
@@ -339,7 +339,7 @@ describe("many-to-many", function() {
     });
 
     describe("cascade updates should be executed when cascadeUpdate option is set", function() {
-        let newPost: Post, newImage: PostImage, savedImage: PostImage;
+        let newPost: Post, newImage: PostImage;
 
         before(reloadDatabase);
 
@@ -355,7 +355,6 @@ describe("many-to-many", function() {
             return postImageRepository
                 .save(newImage)
                 .then(image => {
-                    savedImage = image as PostImage;
                     newPost.images = [];
                     newPost.images.push(image as PostImage);
                     return postRepository.save(newPost);
@@ -389,7 +388,7 @@ describe("many-to-many", function() {
     });
 
     describe("cascade remove should be executed when cascadeRemove option is set", function() {
-        let newPost: Post, newMetadata: PostMetadata, savedMetadata: PostMetadata;
+        let newPost: Post, newMetadata: PostMetadata;
 
         before(reloadDatabase);
 
@@ -405,7 +404,6 @@ describe("many-to-many", function() {
             return postMetadataRepository
                 .save(newMetadata)
                 .then(metadata => {
-                    savedMetadata = metadata as PostMetadata;
                     newPost.metadatas = [];
                     newPost.metadatas.push(metadata as PostMetadata);
                     return postRepository.save(newPost);