+++ CHANGELOG.md
@@ -40,8 +40,10 @@ Use `findOne(id)` method instead now.
 * sqlite now supports relative database file paths (#798 and #799)
 * fixed bug with not properly working `update` method (#1037, #1042)
 * fixed bug with replication support (#1035)
-* fixed bug with wrong embedded column names being generated (#969)
 * fixed broken `typeorm version` command
+* fixed bug with wrong embedded column names being generated (#969) 
+* added support for caching in respositories [#1057](https://github.com/typeorm/typeorm/issues/1057)
+* added support for the `citext` column type for postgres [#1075](https://github.com/typeorm/typeorm/pull/1075)
 
 ## 0.1.1 (current)